+++ crates/pluot_core/src/layers/mod.rs
@@ -1,4 +1,3 @@
-pub mod core;
 pub mod position_utils;
 pub mod point_layer;
 pub mod line_layer;
