+++ crates/pluot_core/src/plots/mod.rs
@@ -1,4 +1,3 @@
-pub mod layered_plot;
 // TODO: remove the following
 //pub mod barplot;
 //pub mod bioimage;
